requirements.txt
aiogram
aiohttp
python-dotenv
apscheduler
pydantic
sqlalchemy
asyncpg
aiosqlite
